refactor: Merge testing module with only consumer
