+++ settings.gradle
@@ -42,4 +42,3 @@ include 'doi-proxy'
 include 'doi-transformer'
 include 'publication-from-doi'
 include 'metadata-transformer'
-include 'fetch-doi-testing'
